% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_significant.R
\name{.get_corrected_thresh}
\alias{.get_corrected_thresh}
\title{Get corrected significance threshold for state}
\usage{
.get_corrected_thresh(x, thresh, p.adjust.method)
}
\arguments{
\item{x}{One column from a \linkS4class{multiStateQTLExperiment} object.}

\item{thresh}{Significance threshold}

\item{p.adjust.method}{Method of multiple-test correction if mode != simple}
}
\description{
Get corrected significance threshold for state
}
