% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_significant.R
\name{.get_thresh_per_state}
\alias{.get_thresh_per_state}
\title{Get the threshold for each state}
\usage{
.get_thresh_per_state(object, mode, thresh, p.adjust.method)
}
\arguments{
\item{object}{A \linkS4class{multiStateQTLExperiment} object.}

\item{mode}{Method to determine significance threshold per state}

\item{thresh}{Significance threshold}

\item{p.adjust.method}{Method of multiple-test correction if mode != simple}
}
\description{
Get the threshold for each state
}
