% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_id.R
\name{feature_id}
\alias{feature_id}
\alias{feature_id,multiStateQTLExperiment-method}
\alias{feature_id<-,multiStateQTLExperiment-method}
\title{Feature ID methods}
\usage{
\S4method{feature_id}{multiStateQTLExperiment}(object, onAbsence = "warn")

\S4method{feature_id}{multiStateQTLExperiment}(object, ...) <- value
}
\arguments{
\item{object}{A \linkS4class{multiStateQTLExperiment} object.}

\item{onAbsence}{String indicating an additional action to take when size factors are absent:
nothing (\code{"none"}), a warning (\code{"warn"}) or an error (\code{"error"}).}

\item{...}{Additional arguments, currently ignored.}

\item{value}{A numeric vector of length equal to \code{nrow(object)}, containing the name of the feature (e.g. gene) being tested.}
}
\value{
For \code{feature_id}, a vector is returned containing the name of the feature tested in each association.

For \code{feature_id<-}, a modified \code{object} is returned with the updated feature_ids in \code{\link{rowData}} and \code{\link{int_rowData}}.
}
\description{
Gets or sets the feature_id for all tests in a \linkS4class{multiStateQTLExperiment} object.
}
\details{
Every QTL refers to the association between a variant and a feature. The \code{feature_id} methods can be used to get or set feature IDs for all tests in a multiStateQTLExperiment object. The values are stored in the \code{\link{rowData}} and in the \code{\link{int_rowData}} as the \code{\link{feature_id}} field so it can be easily accessed but not removed or accidentally overwritten.
}
\examples{
msqe <- mockMSQE()
feature_id(msqe) <- sample(LETTERS, nrow(msqe), replace=TRUE)
feature_id(msqe)

}
\seealso{
\linkS4class{multiStateQTLExperiment}, for the underlying class definition.
}
\author{
Christina B Azodi
}
