% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_load.R
\name{fillNAs}
\alias{fillNAs}
\title{Return MSQE object with NAs filled in}
\usage{
fillNAs(object, verbose)
}
\arguments{
\item{object}{Named array or data.frame with state name and the file to the
QTL summary statistics for that state. If data.frame is
provided, additional columns will be stored in the colData
annotation.}

\item{verbose}{logical. Whether to print progress messages.}
}
\description{
Return MSQE object with NAs filled in
}
