% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_load.R
\name{getRandomQTL}
\alias{getRandomQTL}
\title{Return an array of N random QTL}
\usage{
getRandomQTL(
  input,
  nRandom = nRandom,
  feature = feature,
  snp = snp,
  verbose = verbose
)
}
\description{
Return an array of N random QTL
}
