% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_significant.R
\name{getSignificant}
\alias{getSignificant}
\alias{getSignificant,multiStateQTLExperiment-method}
\title{Add assay with T/F significance calls}
\usage{
getSignificant(object, ...)

\S4method{getSignificant}{multiStateQTLExperiment}(
  object,
  thresh = 0.05,
  second.thresh = thresh,
  feature = "feature_id",
  significance = "pval",
  mode = c("feature-wise-FDR", "simple", "global-FDR"),
  p.adjust.method = c("fdr", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY",
    "none")
)
}
\arguments{
\item{object}{A \linkS4class{multiStateQTLExperiment} object.}

\item{thresh}{Significance threshold}

\item{second.thresh}{Significance threshold for associations with significance in one state.}

\item{feature}{rowData column name with feature identifier}

\item{significance}{assay name with significance score.}

\item{mode}{Method to determine significance threshold per state}

\item{p.adjust.method}{Method of multiple-test correction if mode != simple}
}
\description{
Add assay with T/F significance calls
}
