% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_load.R
\name{getSignificantQTL}
\alias{getSignificantQTL}
\title{Return an array of the QTL significant in at least one state}
\usage{
getSignificantQTL(
  input,
  feature = feature,
  snp = snp,
  sig = sig,
  thresh = thresh,
  verbose = verbose
)
}
\description{
Return an array of the QTL significant in at least one state
}
