% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_load.R
\name{getTopQTL}
\alias{getTopQTL}
\title{Return an array of the top QTL for each feature across all states}
\usage{
getTopQTL(input, feature = feature, snp = snp, sig = sig, verbose = verbose)
}
\description{
Return an array of the top QTL for each feature across all states
}
