% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_load.R
\name{loadSummaryStatistics}
\alias{loadSummaryStatistics}
\title{Generate multiStateQTLExperiment object from QTL mapping results}
\usage{
loadSummaryStatistics(
  input,
  mode = "significant",
  na.rm = FALSE,
  thresh = 0.05,
  nRandom = 1000,
  feature = "feature_id",
  snp = "snp_id",
  beta = "beta",
  se = "beta_se",
  sig = "empirical_feature_p_value",
  verbose = TRUE
)
}
\arguments{
\item{input}{Named array or data.frame with state name and the file to the
QTL summary statistics for that state. If data.frame is
provided, additional columns will be stored in the colData
annotation.}

\item{mode}{How to extract summary statistics.}

\item{na.rm}{logical. If QTL should be dropped if missing for any state.}

\item{thresh}{Significance threshold if `mode=significant`.}

\item{nRandom}{Number of QTL to sample if `mode=random`.}

\item{feature}{The name/index of the column with the feature ID.}

\item{snp}{The name/index of the column with the SNP ID.}

\item{beta}{The name/index of the column with the effect size/beta value.}

\item{se}{The name/index of the column with the effect size/beta standard
error value.}

\item{sig}{The name/index of the column with the significance score.}

\item{verbose}{logical. Whether to print progress messages.}
}
\description{
A suite of methods to extract QTL mapping summary statistics from common
QTL workflow output files.
}
\section{Modes}{


\describe{
\item{\code{mode=all}:}{Loads all QTL in input data.}
\item{\code{mode=significant}:}{Loads all QTL under the significance
threshold in at least one state.}
\item{\code{mode=top}:}{Loads all QTL that are the top QTL for a feature in
at least one state.}
\item{\code{mode=random}:}{Loads a random subset of QTL (n=`nRandom`).}
}
}

\author{
Christina B Azodi
}
