% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainExpName.R
\docType{methods}
\name{mainExpName}
\alias{mainExpName}
\alias{mainExpName<-}
\alias{mainExpName,multiStateQTLExperiment-method}
\alias{mainExpName<-,multiStateQTLExperiment,character_OR_NULL-method}
\title{Get or set object experiment name}
\usage{
\S4method{mainExpName}{multiStateQTLExperiment}(x)

\S4method{mainExpName}{multiStateQTLExperiment,character_OR_NULL}(x) <- value
}
\arguments{
\item{x}{A \linkS4class{multiStateQTLExperiment} object.}

\item{value}{Any character-like object or \code{NULL} to remove existing name.}
}
\value{
For \code{mainExpName}, a character object is returned containing the object
name. If no name is available, a \code{NULL} is returned.

For \code{mainExpName<-}, a modified \code{x} is returned with a name in
\code{\link{mainExpName}}.
}
\description{
Get or set mainExpName for a \linkS4class{multiStateQTLExperiment} instance.
}
\examples{
msqe <- mockMSQE()
mainExpName(msqe) <- "demo_experiment"
mainExpName(msqe)

}
\seealso{
\linkS4class{multiStateQTLExperiment}, for the underlying class definition.
}
\author{
Christina Azodi
}
