% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash_2_msqe.R
\name{mash_2_msqe}
\alias{mash_2_msqe}
\alias{.mashData_2_msqe}
\alias{.mashFit_2_msqe}
\title{Coercing mash data objects into MSQE objects}
\usage{
mash_2_msqe(data, sep = NULL, rowData = NULL, verbose = FALSE)

.mashData_2_msqe(data)

.mashFit_2_msqe(data)
}
\arguments{
\item{data}{A `MSQE` object or named list containing "betas" and "error"
matricies.}

\item{verbose}{Logical.}

\item{thresh}{Threshold for calling tests as significant (pval or lfsr).}
}
\description{
Coercing mash data objects into MSQE objects
}
\examples{

}
