% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_load.R
\name{mash_set_data2MSQE}
\alias{mash_set_data2MSQE}
\title{Convert a mash_set_data object to a MSQE}
\usage{
mash_set_data2MSQE(m)
}
\arguments{
\item{m}{mash_set_data input}
}
\description{
Convert a mash_set_data object to a MSQE
}
