% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtle_misc.R
\docType{methods}
\name{QTLe-miscellaneous}
\alias{QTLe-miscellaneous}
\alias{objectVersion}
\alias{objectVersion,QTLExperiment-method}
\alias{mainExpName}
\alias{mainExpName,QTLExperiment-method}
\alias{mainExpName<-}
\alias{mainExpName<-,QTLExperiment,character_OR_NULL-method}
\title{Miscellaneous methods to apply to a
\linkS4class{QTLExperiment} object.}
\arguments{
\item{x}{A \linkS4class{QTLExperiment} object.}

\item{value}{Any character-like object or \code{NULL} to
remove existing labels.}

\item{...}{Additional arguments.}
}
\description{
Miscellaneous methods for the \linkS4class{QTLExperiment} class.
}
\section{Available methods}{

In the following code snippets, \code{x} is a
\linkS4class{QTLExperiment} objects.

\describe{
\item{\code{objectVersion(x)}:}{Return the version of the package with which
                                \code{x} was constructed.}
\item{\code{mainExpName(x)}:}{Return the name assigned to the \code{x}.}
}
}

\examples{
qtle <- mockQTLE()
objectVersion(qtle)
mainExpName(qtle)
mainExpName(qtle) <- "test_name"
mainExpName(qtle)

}
\seealso{
\linkS4class{QTLExperiment}, for the underlying class definition
and \code{\link{updateObject}} when \code{objectVersion} is used.
}
\author{
Christina B. Azodi
}
