% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockData.R
\name{mockMSQE}
\alias{mockMSQE}
\title{Mock MSQE object}
\usage{
mockMSQE(nStates = 10, nQTL = 100, seed = NULL)
}
\arguments{
\item{nStates}{Number of states}

\item{nQTL}{Number of QTL associations}
}
\description{
Mock MSQE object
}
