% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/multiStateQTLExperiment.R
\docType{class}
\name{multiStateQTLExperiment-class}
\alias{multiStateQTLExperiment-class}
\alias{multiStateQTLExperiment}
\alias{coerce,SummarizedExperiment,multiStateQTLExperiment-method}
\alias{coerce,RangedSummarizedExperiment,multiStateQTLExperiment-method}
\title{The multiStateQTLExperiment class}
\usage{
multiStateQTLExperiment(..., reducedDims = list())
}
\arguments{
\item{...}{Arguments passed to the \code{\link{SummarizedExperiment}}
constructor to fill the slots of the base class.}

\item{reducedDims}{A list of matrix-like objects containing dimensionality
reduction results, each of which should have the same number of rows as the
output multiStateQTLExperiment object.}
}
\value{
A multiStateQTLExperiment object.
}
\description{
The multiStateQTLExperiment class is designed to represent multi-state QTL
data. It inherits from the \linkS4class{RangedSummarizedExperiment} class.
In addition, the class supports storage of multi-state adjusted
beta and betaSE results (e.g., mash) and
storage of summary results (e.g., pairwise sharing).
}
\details{
In this class, rows should represent associations (feature_id:variant_id pairs)
while columns represent states (e.g. cell types). Assays include betas and
error associated with the betas (e.g. standard errors).
As with any \linkS4class{SummarizedExperiment} derivative,
different quantifications (e.g., counts, CPMs, log-expression) can be stored
simultaneously in the \code{\link{assays}} slot,
and row and column metadata can be attached using \code{\link{rowData}} and
 \code{\link{colData}}, respectively.

The extra arguments in the constructor (e.g., \code{\link{reducedDims}})
represent the main extensions implemented in the SingleCellExperiment class.
This enables a consistent, formalized representation of data structures
that are commonly encountered during single-cell data analysis.
Readers are referred to the specific documentation pages for more details.

A MSQE can also be coerced from a \linkS4class{SummarizedExperiment} or
\linkS4class{RangedSummarizedExperiment} instance.
}
\examples{
nStates <- 10
nQTL <- 100
betas <- matrix(rnorm(nStates * nQTL), ncol=nStates)
error <- matrix(abs(rnorm(nStates * nQTL)), ncol=nStates)
rowdata <- DataFrame(feature_id = sample(1:10, nQTL, replace=TRUE),
                     variant_id = sample(seq(1e3:1e5), nQTL))

msqe <- multiStateQTLExperiment(assays=list(betas=betas, error=error),
                                rowData=rowdata)
msqe

## coercion from SummarizedExperiment
se <- SummarizedExperiment(assays=list(betas=betas, error=error),
                           rowData=rowdata)
as(se, "multiStateQTLExperiment")

}
\author{
Christina B Azodi
}
