% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_smart-subset.R
\name{naReplace}
\alias{naReplace}
\title{Return MSQE object with NAs filled in}
\usage{
naReplace(object, verbose = FALSE)
}
\arguments{
\item{object}{`MSQE` object}

\item{verbose}{logical. Whether to print progress messages.}
}
\description{
Return MSQE object with NAs filled in
}
