% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtle_recover_ids.R
\name{QTLe-recover}
\alias{QTLe-recover}
\alias{.recover_qtle_ids}
\title{Recover QTLExperiment IDs}
\usage{
.recover_qtle_ids(object)
}
\arguments{
\item{object}{QTLExperiment object}
}
\value{
A \linkS4class{QTLExperiment} object with recovered rowData or colData.
}
\description{
Function to recover protected rowData (feature_id, variant_id)
and colData (state_id) from internal structure.
}
