% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top-per-feature.R
\name{subsetPerFeature}
\alias{subsetPerFeature}
\title{Function to subset get top QTL for each feature}
\usage{
subsetPerFeature(object, assay = "pval", FUN = which.min)
}
\arguments{
\item{object}{A \linkS4class{multiStateQTLExperiment} object with multiple
QTL tests (i.e., rows) for at least one feature.}

\item{assay}{Assay to use as filter assay}

\item{feature}{Name of feature id in colData(object).}

\item{keep}{min/max to keep the test for each feature with the min or the
max value in the assay.}
}
\description{
This function pulls the top QTL for each feature (e.g., gene) and
outputs a **new** \linkS4class{multiStateQTLExperiment}. Note that the
output from this function should be assigned to a new variable as it will
overwrite if assigned to the same variable.
}
