% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_smart-subset.R
\name{subsetSignificant}
\alias{subsetSignificant}
\title{Return an array of the QTL significant in at least one state}
\usage{
subsetSignificant(msqe, assay = "pval", thresh = thresh, verbose = FALSE)
}
\arguments{
\item{msqe}{`MSQE` object}

\item{assay}{The name of the assay containing significance information.}

\item{thresh}{Significance threshold.}

\item{verbose}{logical. Whether to print progress messages.}
}
\description{
Return an array of the QTL significant in at least one state
}
