% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_load.R
\name{summaryStats_2_msqe}
\alias{summaryStats_2_msqe}
\title{Coerce QTL summary statistics into a multiStateQTLExperiment object}
\usage{
summaryStats_2_msqe(
  input,
  feature_id = "gene_id",
  variant_id = "variant_pos",
  betas = "slope",
  error = "slope_se",
  pval = "pval_perm",
  n_max = Inf,
  verbose = TRUE
)
}
\arguments{
\item{input}{Named array or data.frame with state name and the file to the
QTL summary statistics for that state. If data.frame is
provided, additional columns will be stored in the colData
annotation.}

\item{feature_id}{The name/index of the column with the feature_id info.}

\item{variant_id}{The name/index of the column with the variant_id info.}

\item{betas}{The name/index of the column with the effect size/beta value.}

\item{error}{The name/index of the column with the effect size/beta standard
error value.}

\item{pval}{The name/index of the column with the significance score.}

\item{n_max}{Max number of rows to read per file. This is primarily used
for testing purposes.}

\item{verbose}{logical. Whether to print progress messages.}
}
\description{
A suite of methods to extract QTL mapping summary statistics from common
QTL workflow output files.
}
\author{
Christina B Azodi
}
