% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_assays.R
\docType{methods}
\name{msqe-assays}
\alias{msqe-assays}
\alias{betas}
\alias{betas<-}
\alias{betas,multiStateQTLExperiment-method}
\alias{betas<-,multiStateQTLExperiment-method}
\alias{error}
\alias{error<-}
\alias{error,multiStateQTLExperiment-method}
\alias{error<-,multiStateQTLExperiment-method}
\alias{pval}
\alias{pval<-}
\alias{pval,multiStateQTLExperiment-method}
\alias{pval<-,multiStateQTLExperiment-method}
\alias{lfsr}
\alias{lfsr<-}
\alias{lfsr,multiStateQTLExperiment-method}
\alias{lfsr<-,multiStateQTLExperiment-method}
\title{Named assay getters and setters}
\description{
These are methods for getting or setting \code{assay(msqe, i=X, ...)}
where \code{msqe} is a \linkS4class{multiStateQTLExperiment} object and
\code{X} is the name of the method. For example, \code{betas} will get or
set \code{X="betas"}.
}
\section{Available methods}{

Here \code{x} is a \linkS4class{multiStateQTLExperiment} object,
\code{value} is a matrix-like object with the same dimensions as \code{x},
and \code{...} are further arguments passed to \code{\link{assay}}
(for the getter) or \code{\link{assay<-}} (for the setter).

\describe{
\item{\code{betas(x, ...)}, \code{betas(x, ...) <- value}:}{
Get or set a matrix of raw betas, i.e., QTL effect sizes.
}
\item{\code{se(x, ...)}, \code{se(x, ...) <- value}:}{
Get or set a matrix of raw beta standard errors.
}
\item{\code{sig(x, ...)}, \code{sig(x, ...) <- value}:}{
Get or set a matrix of raw significance scores (e.g. pvals, qvals)
}
\item{\code{adj_betas(x, ...)}, \code{adj_betas(x, ...) <- value}:}{
Get or set a matrix of multi-state adjusted betas, i.e., QTL effect sizes
adjusted by mashr.
}
\item{\code{adj_se(x, ...)}, \code{adj_se(x, ...) <- value}:}{
Get or set a matrix of multi-state adjusted beta standard errors.
}
\item{\code{adj_sigl(x, ...)}, \code{adj_sigl(x, ...) <- value}:}{
Get or set a matrix of multi-state adjusted significance scores (e.g. lfsr
output from mashr)
}
}
}

\examples{
msqe <- mockMSQE()
new_betas <- matrix(rnorm(nrow(msqe)*ncol(msqe)), ncol=ncol(msqe))
row.names(new_betas) <- row.names(msqe)
colnames(new_betas) <- colnames(msqe)
betas(msqe) <- new_betas
dim(betas(msqe))

}
\seealso{
\code{\link{assay}} and \code{\link{assay<-}}, for the wrapped methods.
}
\author{
Christina B Azodi
}
