% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_combine.R
\docType{methods}
\name{MSQE-combine}
\alias{MSQE-combine}
\alias{cbind,multiStateQTLExperiment-method}
\alias{rbind,multiStateQTLExperiment-method}
\title{Combining multiStateQTLExperiment objects}
\description{
An overview of methods to combine multiple
\linkS4class{multiStateQTLExperiment} objects by row or column.
These methods ensure that all data fields remain synchronized when states
or associations are added or removed.
}
\section{Combining}{

In the following examples, \code{...} contains one or more
\linkS4class{multiStateQTLExperiment} object.
\describe{
\item{\code{rbind(..., deparse.level=1)}:}{Returns a
\linkS4class{multiStateQTLExperiment} object where all objects are combined
row-wise. All objects in \code{...} must have the same
values for \code{\link{reducedDims}}. Metadata is combined as in
\code{?"\link{rbind,SummarizedExperiment-method}"}. The \code{deparse.level}
specifies how row.names are generated as described in
\code{?\link[base]{rbind}}.
}

\item{\code{cbind(..., deparse.level=1)}:}{Returns a
\linkS4class{multiStateQTLExperiment} object where all objects are combined
column-wise. All objects in \code{...} must have the exact same values for
\code{\link{reducedDims}}, however they can be unordered. Dimensionality
reduction results with the same name across objects will be combined row-wise
to create the corresponding entry in the output object. Metadata is combined
as in \code{?"\link{cbind,SummarizedExperiment-method}"}. The
\code{deparse.level} specifies how colnames are generated as described in
\code{?\link[base]{cbind}}.
}
}
}

\examples{
msqe <- mockMSQE()

# Combining:
rbind(msqe, msqe)
cbind(msqe, msqe)

}
\author{
Christina B Azodi
}
