% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_combine-subset.R
\docType{methods}
\name{MSQE-combine}
\alias{MSQE-combine}
\alias{cbind,multiStateQTLExperiment-method}
\alias{rbind,multiStateQTLExperiment-method}
\alias{[,multiStateQTLExperiment,ANY-method}
\alias{[,multiStateQTLExperiment,ANY,ANY-method}
\alias{[,multiStateQTLExperiment,ANY,ANY,ANY-method}
\alias{[<-,multiStateQTLExperiment,ANY,ANY,multiStateQTLExperiment-method}
\title{Combining or subsetting multiStateQTLExperiment objects}
\description{
An overview of methods to combine multiple \linkS4class{multiStateQTLExperiment}
objects by row or column, or to subset a multiStateQTLExperiment by row or
column. These methods ensure that all data fields remain synchronized when
states or associations are added or removed.
}
\section{Combining}{

In the following code snippets, \code{...} contains one or more
\linkS4class{multiStateQTLExperiment} objects.
\describe{
\item{\code{rbind(..., deparse.level=1)}:}{Returns a multiStateQTLExperiment 
where all objects are combined row-wise.

Note that all objects in \code{...} must have the exact same values for
\code{\link{reducedDims}}.

Metadata is combined as in \code{?"\link{rbind,SummarizedExperiment-method}"}.
The \code{deparse.level} specifies how row.names are generated as described 
in \code{?\link[base]{rbind}}.
}

\item{\code{cbind(..., deparse.level=1)}:}{Returns a multiStateQTLExperiment
where all objects are combined column-wise.

Note that all objects in \code{...} must have the exact same values for
\code{\link{reducedDims}} (though they can be unordered). Dimensionality 
reduction results with the same name across objects will be combined row-wise
to create the corresponding entry in the output object.

Metadata is combined as in \code{?"\link{cbind,SummarizedExperiment-method}"}.
The \code{deparse.level} specifies how colnames are generated as described 
in \code{?\link[base]{cbind}}.
}
}
}

\section{Subsetting}{

In the following, \code{x} is a \linkS4class{multiStateQTLExperiment} object.

\describe{
\item{\code{x[i, j, ..., drop=TRUE]}:}{Returns a multiStateQTLExperiment
 containing the specified rows \code{i} and columns \code{j}.

\code{i} and \code{j} can be a logical, integer or character vector of
subscripts, indicating the rows and columns respectively to retain. If either
is missing, subsetting is only performed in the specified dimension.

Arguments in \code{...} and \code{drop} are passed to
\code{\link{[,SummarizedExperiment-method}}.}

\item{\code{x[i, j, ...] <- value}:}{Replaces all data for rows \code{i} and
columns {j} with the corresponding fields in a multiStateQTLExperiment
\code{value}.

\code{i} and \code{j} can be a logical, integer or character vector of
subscripts, indicating the rows and columns respectively to retain. If either
is missing, subsetting is only performed in the specified dimension.
If both are missing, \code{x} is replaced entirely with \code{value}.

Arguments in \code{...} are passed to the corresponding
\linkS4class{SummarizedExperiment} method.}
}
}

\examples{
msqe <- mockMSQE()

# Combining:
rbind(msqe, msqe)
cbind(msqe, msqe)

# Subsetting:
msqe[1:10,]
msqe[,1:5]

msqe2 <- msqe
msqe2[1:10,] <- msqe[11:20,]

# Can also use subset()
sce$WHEE <- sample(LETTERS, ncol(sce), replace=TRUE)
subset(sce, , WHEE=="A")

# Can also use split()
split(sce, sample(LETTERS, nrow(sce), replace=TRUE))

}
\author{
Christina B Azodi
}
