% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_load.R
\name{getRandomQTL}
\alias{getRandomQTL}
\title{Return an array of N random QTL}
\usage{
getRandomQTL(input, nRandom = nRandom, verbose = verbose)
}
\arguments{
\item{input}{Named array or data.frame with state name and the file to the
QTL summary statistics for that state. If data.frame is
provided, additional columns will be stored in the colData
annotation.}

\item{nRandom}{Number of QTL to sample if `mode=random`.}

\item{verbose}{logical. Whether to print progress messages.}
}
\description{
Return an array of N random QTL
}
