% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash_2_msqe.R
\name{mash_2_msqe}
\alias{mash_2_msqe}
\alias{.mashData_2_msqe}
\alias{.mashFit_2_msqe}
\title{Coercing mash data objects into MSQE objects}
\usage{
mash_2_msqe(data, sep = NULL, rowData = NULL, verbose = FALSE)

.mashData_2_msqe(data)

.mashFit_2_msqe(data)
}
\arguments{
\item{data}{A mashr object output from mashr::mash_set_data or mashr::mash.}

\item{sep}{String separating the feature_id from the variant_id in the row.names of the mashr object}

\item{rowData}{if feature_id and variant_id are not in the row.names, a rowData matrix can be provided with this information.}

\item{verbose}{Logical.}
}
\description{
Coercing mash data objects into MSQE objects
}
