% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_misc.R
\docType{methods}
\name{MSQE-miscellaneous}
\alias{MSQE-miscellaneous}
\alias{objectVersion}
\alias{objectVersion,multiStateQTLExperiment-method}
\alias{mainExpName}
\alias{mainExpName,multiStateQTLExperiment-method}
\alias{mainExpName<-}
\alias{mainExpName<-,multiStateQTLExperiment-method}
\alias{colLabels}
\alias{colLabels,multiStateQTLExperiment-method}
\title{Miscellaneous multiStateQTLExperiment methods}
\arguments{
\item{x}{A \linkS4class{multiStateQTLExperiment} object.}

\item{value}{For `colLabels()`, any vector-like object of \code{ncol(object)}
length containing labels for all states. Or \code{NULL} to
remove existing labels.}

\item{...}{Additional arguments.}

\item{onAbsence}{For `colLabels()` a string indicating an additional action
to take when labels are absent: nothing (\code{"none"}),
a warning (\code{"warn"}) or an error (\code{"error"}).}
}
\description{
Miscellaneous methods for the \linkS4class{multiStateQTLExperiment} class.
}
\section{Available methods}{

In the following code snippets, \code{x} and \code{object} are
\linkS4class{multiStateQTLExperiment} objects.
\describe{
\item{\code{show(object)}:}{Print a message to screen describing the
                            contents of \code{object}.}
\item{\code{objectVersion(x)}:}{Return the version of the package with which
                                \code{x} was constructed.}
\item{\code{mainExpName(x)}:}{Return the name assigned to the \code{x}.}
\item{\code{colLabels(x)}:}{Return an array with state names from \code{x}.}
}
}

\examples{
msqe <- mockMSQE()
objectVersion(msqe)
mainExpName(msqe)
mainExpName(msqe) <- "test_name"
mainExpName(msqe)
colLabels(msqe) <- sample(LETTERS, ncol(msqe), replace=TRUE)
colLabels(msqe)

}
\seealso{
\linkS4class{multiStateQTLExperiment}, for the underlying class definition
and \code{\link{updateObject}} when \code{objectVersion} is used.
}
\author{
Christina Azodi
}
