% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_misc.R
\docType{methods}
\name{SCE-miscellaneous}
\alias{SCE-miscellaneous}
\alias{objectVersion}
\alias{objectVersion,multiStateQTLExperiment-method}
\title{Miscellaneous multiStateQTLExperiment methods}
\description{
Miscellaneous methods for the \linkS4class{multiStateQTLExperiment} class
that do not fit in any other documentation category.
}
\section{Available methods}{

In the following code snippets, \code{x} and \code{object} are \linkS4class{multiStateQTLExperiment} objects.
\describe{
\item{\code{show(object)}:}{Print a message to screen describing the contents of \code{object}.}
\item{\code{objectVersion(x)}:}{Return the version of the package with which \code{x} was constructed.}
}
}

\examples{
example(multiStateQTLExperiment, echo=FALSE) # Using the class example
objectVersion(sce)

}
\seealso{
\code{\link{updateObject}}, where \code{objectVersion} is used.
}
\author{
Christina Azodi
}
