% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockData.R
\docType{methods}
\name{mockQTLE}
\alias{mockQTLE}
\alias{mockSummaryStats}
\alias{mockMASHR}
\alias{mockMASHR_FIT}
\title{Mock data for the QTLExperiment object}
\usage{
mockQTLE(nStates = 10, nQTL = 100, names = TRUE)

mockSummaryStats(nStates = 10, nQTL = 100, names = TRUE)

mockMASHR(nStates = 10, nQTL = 100)

mockMASHR_FIT(nStates = 10, nQTL = 100)
}
\arguments{
\item{nStates}{Number of states}

\item{nQTL}{Number of QTL associations}

\item{names}{Logical to include column and row names}
}
\description{
Functions to create fake input data for QTLExperiments.
}
\author{
Christina B Azodi
}
