% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/multiStateQTLExperiment.R
\docType{class}
\name{multiStateQTLExperiment-class}
\alias{multiStateQTLExperiment-class}
\alias{multiStateQTLExperiment}
\alias{coerce,SummarizedExperiment,multiStateQTLExperiment-method}
\alias{coerce,RangedSummarizedExperiment,multiStateQTLExperiment-method}
\title{The multiStateQTLExperiment class}
\usage{
multiStateQTLExperiment(..., reducedDims = list())
}
\arguments{
\item{...}{Arguments passed to the \code{\link{SummarizedExperiment}} constructor to fill the slots of the base class.}
}
\value{
A multiStateQTLExperiment object.
}
\description{
The multiStateQTLExperiment class is designed to represent multi-state QTL
data. It inherits from the \linkS4class{RangedSummarizedExperiment} class.
In addition, the class supports storage of multi-state adjusted
beta and betaSE results (e.g., mash) and
storage of summary results (e.g., pairwise sharing) via
\code{\link{getPairwiseSharing}}.
}
\details{
In this class, rows should represent associations (feature:variant pairs)
while columns represent states (e.g. cell types). Assays include beta and betaSE.
As with any \linkS4class{SummarizedExperiment} derivative,
different quantifications (e.g., counts, CPMs, log-expression) can be stored
simultaneously in the \code{\link{assays}} slot,
and row and column metadata can be attached using \code{\link{rowData}} and
 \code{\link{colData}}, respectively.

The extra arguments in the constructor (e.g., \code{\link{reducedDims}})
represent the main extensions implemented in the SingleCellExperiment class.
This enables a consistent, formalized representation of data structures
that are commonly encountered during single-cell data analysis.
Readers are referred to the specific documentation pages for more details.

A MSQE can also be coerced from a \linkS4class{SummarizedExperiment} or
\linkS4class{RangedSummarizedExperiment} instance.
}
\examples{
nStates <- 10
nQTL <- 100
b <- matrix(rnorm(nStates * nQTL), ncol=nStates)
se <- matrix(abs(rnorm(nStates * nQTL)), ncol=nStates)

pca <- matrix(runif(nStates*5), nStates)

msqe <- multiStateQTLExperiment(assays=list(betas=b, se=se),
                                reducedDims=SimpleList(PCA=pca))
msqe

## coercion from SummarizedExperiment
se <- SummarizedExperiment(assays=list(betas=b, se=se))
as(se, "multiStateQTLExperiment")

}
\author{
Christina B Azodi
}
