% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqe_subset.R
\docType{methods}
\name{MSQE-subset}
\alias{MSQE-subset}
\alias{[,multiStateQTLExperiment,ANY-method}
\alias{[,multiStateQTLExperiment,ANY,ANY-method}
\alias{[,multiStateQTLExperiment,ANY,ANY,ANY-method}
\alias{[<-,multiStateQTLExperiment,ANY,ANY,multiStateQTLExperiment-method}
\title{Subsetting and replacing data in multiStateQTLExperiment objects}
\description{
An overview of methods to subset a
\linkS4class{multiStateQTLExperiment} object by row and/or column, and to
replace all data for the specified rows and/or columns with another value.
These methods ensure that all data fields remain synchronized when states
or associations are removed.
}
\section{Subsetting}{

In the following, \code{x} is a \linkS4class{multiStateQTLExperiment} object.

\describe{
\item{\code{x[i, j, ..., drop=TRUE]}:}{Returns a multiStateQTLExperiment
 containing the specified rows \code{i} and columns \code{j}, where \code{i}
 and \code{j} can be a logical, integer or character vector of subscripts,
 indicating the rows and columns, respectively, to retain. If either is
 missing, subsetting is only performed in the specified dimension. Arguments
 in \code{...} and \code{drop} are passed to
 \code{\link{[,SummarizedExperiment-method}}.}
}
}

\section{Replacing}{

In the following, \code{x} is a \linkS4class{multiStateQTLExperiment} object.

\describe{
\item{\code{x[i, j, ...] <- value}:}{Replaces all data for rows \code{i} and
columns {j} with the corresponding fields in a multiStateQTLExperiment
\code{value}, where \code{i} and \code{j} can be a logical, integer, or
character vector of subscripts, indicating the rows and columns,
respectively, to retain. If either is missing, subsetting is only performed
in the specified dimension. If both are missing, \code{x} is replaced
entirely with \code{value}. Arguments in \code{...} are passed to the
corresponding \linkS4class{SummarizedExperiment} method.}
}
}

\examples{
msqe <- mockMSQE()

# Subsetting:
msqe[1:10,]
msqe[,1:5]

msqe2 <- msqe
msqe2[1:10,] <- msqe[11:20,]

# Can also use subset()
sce$WHEE <- sample(LETTERS, ncol(sce), replace=TRUE)
subset(sce, , WHEE=="A")

# Can also use split()
split(sce, sample(LETTERS, nrow(sce), replace=TRUE))

}
\author{
Christina B Azodi
}
