% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtle_version.R
\docType{methods}
\name{QTLe-version}
\alias{QTLe-version}
\alias{objectVersion}
\alias{objectVersion,QTLExperiment-method}
\title{Return the version of a \linkS4class{QTLExperiment} object}
\arguments{
\item{x}{A \linkS4class{QTLExperiment} object.}
}
\value{
A package version, of class \link{package_version}.
}
\description{
Specifies the version of the \link{QTLExperiment} package that an object of class
\linkS4class{QTLExperiment} was created with.
}
\section{Available methods}{

In the following code snippets, \code{x} is a
\linkS4class{QTLExperiment} objects.

\describe{
\item{\code{objectVersion(x)}:}{Return the version of the package with which
                                \code{x} was constructed.}
 }
}

\examples{
qtle <- mockQTLE()
objectVersion(qtle)

}
\seealso{
\linkS4class{QTLExperiment}, for the underlying class definition
and \code{\link{updateObject}} to update the object to the latest version.
}
\author{
Christina B. Azodi, Amelia Dunstone
}
